nightly-build: use the storage-gluster-common buildroot from CentOS CBS
Using an EPEL buildroot for CentOS packages is not correct. This can
cause problems when packages are not matching in the EPEL and CentOS
repositories.

Use the Koji profile 'cbs' provided by the centos-packager RPM to fetch
the mock-config from the CentOS CBS. Use this config to rebuild the
src.rpm so the resulting packages have correct dependencies.

Updates: #22
Signed-off-by: Niels de Vos <ndevos@redhat.com>