heketi-functional: increase timeout to 3 hours
This is a short term fix to deal with the fact that heketi functional tests are regularly timing out now, mainly due to the increased number of functional tests we've added. A longer term project to make the functional tests faster is needed but for now we're just going to increase the timeout.