nightly-builds: install latest centos-release-gluster package
With the update to CentOS 7.1810 it seems that yum changed behaviour a
little. For some (unclear at the moment) reason yum decides to install
centos-release-gluster-legacy instead of centos-release-gluster5.

By installing centos-release-storage-common *before* requesting to
install centos-release-gluster, yum picks the right version.

This is a workaround as installing centos-release-gluster on a cleanly
installed system should get the latest maintained Gluster release.