csi-driver: run smoke test on the master branch after merging changes (#38)
The badge on the gluster-csi-driver project reflects the last run of the
smoke test. These tests are run against pull-requests, and not the
actual merged changes in master.

Introduce a 2nd job that runs the same smoke tests, but only once there
are pushed changes in the master branch. This uses the Jenkins GitHub
plugin as explained on the CentOS Wiki. As the page explains, it is
required to configure the GitHub project with a Webhook pointing to the
CentOS CI as well.

Once this change has been merged, the image in the README can be updated
to https://ci.centos.org/view/Gluster/job/gluster_csi-driver-master

URL: https://wiki.centos.org/QaWiki/CI/GithubIntegration
Updates: gluster/gluster-csi-driver#96