nightly-builds: fix shallow cloning of git repository for non-master branches
When passing `--depth 1` to `git clone`, only the last commit of the
master branch will be fetched. Changing branches afterwards does not
work, as no other branches have been downloaded.

In order to do nightly builds for non-master branches, the `--branch`
argument to `git clone` is required. With this, it is not needed anymore
to checkout the branch afterwards.

Fixes: b552df317a1 ("git clone with just depth 1 for glusterfs")
Signed-off-by: Niels de Vos <ndevos@redhat.com>