nightly-builds: use date of last commit as part of the package version
There is no need to have a new version every day (night). By using the
date of the last commit in the package version, the packages will be
built, but the version/release stays the same. This reduces the number
of packages in the repository, and makes it more obvious what changes
were included. The git-hash is still part of the version as well.

Signed-off-by: Niels de Vos <ndevos@redhat.com>