Adding task to setup-glusto.yml playbook to install crefi.
At present we only use a script called file_dir_ops.py in
glusto-tests to run IO. But the issue with file_dir_ops.py
is that it isn't that effective to create files. A better
tool which is often used is Crefi which can be used to create
a varity of fops.
Link to tool is avaliable here:
https://github.com/vijaykumar-koppad/Crefi

Hence to support running crefi in glustolibs I am adding a task
to setup crefi to setup-glusto.yml.

Signed-off-by: kshithijiyer <kshithij.ki@gmail.com>