Deploy Jenkins jobs with GitHub Action
The new tests/Containerfile builds a container image with the sources
for the Jenking jobs and jenkins-job-builder. By default the container
runs the tests/verify-yaml.sh script. When passing
tests/deploy-centos-ci.sh as alternative entrypoint, the jobs will be
updated in Jenkins.

Use the JENKINS_API_KEY set in the secrets of the GitHub repository and
use it to push the changes to the Jenkins instance in the CentOS CI.

Signed-off-by: Niels de Vos <ndevos@redhat.com>